+++ src/Sentry/Protocol/Mechanism.cs
@@ -29,6 +29,12 @@ public sealed class Mechanism : ISentryJsonSerializable
     /// </summary>
     public static readonly string DescriptionKey = "Sentry:Description";
 
+    /// <summary>
+    /// Key found inside of <c>Exception.Data</c> describing whether the exception is considered terminal.
+    /// </summary>
+    /// <remarks> It's not prefixed with 'Sentry' so the MainExceptionProcessor does not remove it. </remarks>
+    public static readonly string TerminalKey = "Terminal";
+
     internal Dictionary<string, object>? InternalData { get; private set; }
 
     internal Dictionary<string, object>? InternalMeta { get; private set; }